Model CMSG_QUESTGIVER_ACCEPT_QUEST's trailing unk1 for Wrath 3.3.5
The 3.3.5a client sends the quest-accept packet as guid + quest_id + a
trailing u32 (unk1) that AzerothCore reads in HandleQuestgiverAcceptQuest;
the crate modeled it as 12 bytes, so the strict size check rejected every
real packet. Add unk1 for 3.3.5 only; 1.12 and 2.4.3 stay at 12 bytes.
