+++ wow_world_messages/src/world/wrath/cmsg_questgiver_accept_quest.rs
@@ -1,0 +1,114 @@
+use std::io::{Read, Write};
+
+use crate::Guid;
+
+/// The 3.3.5a client appends a trailing u32 that AzerothCore reads as `unk1`.
+/// Auto generated from the original `wowm` in file [`wow_message_parser/wowm/world/quest/cmsg_questgiver_accept_quest.wowm:9`](https://github.com/gtker/wow_messages/tree/main/wow_message_parser/wowm/world/quest/cmsg_questgiver_accept_quest.wowm#L9):
+/// ```text
+/// cmsg CMSG_QUESTGIVER_ACCEPT_QUEST = 0x0189 {
+///     Guid guid;
+///     u32 quest_id;
+///     u32 unk1;
+/// }
+/// ```
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord, Default)]
+pub struct CMSG_QUESTGIVER_ACCEPT_QUEST {
+    pub guid: Guid,
+    pub quest_id: u32,
+    pub unk1: u32,
+}
+
+impl crate::private::Sealed for CMSG_QUESTGIVER_ACCEPT_QUEST {}
+impl CMSG_QUESTGIVER_ACCEPT_QUEST {
+    fn read_inner(mut r: &mut &[u8], body_size: u32) -> Result<Self, crate::errors::ParseErrorKind> {
+        if body_size != 16 {
+            return Err(crate::errors::ParseErrorKind::InvalidSize);
+        }
+
+        // guid: Guid
+        let guid = crate::util::read_guid(&mut r)?;
+
+        // quest_id: u32
+        let quest_id = crate::util::read_u32_le(&mut r)?;
+
+        // unk1: u32
+        let unk1 = crate::util::read_u32_le(&mut r)?;
+
+        Ok(Self {
+            guid,
+            quest_id,
+            unk1,
+        })
+    }
+
+}
+
+impl crate::Message for CMSG_QUESTGIVER_ACCEPT_QUEST {
+    const OPCODE: u32 = 0x0189;
+
+    #[cfg(feature = "print-testcase")]
+    fn message_name(&self) -> &'static str {
+        "CMSG_QUESTGIVER_ACCEPT_QUEST"
+    }
+
+    #[cfg(feature = "print-testcase")]
+    fn to_test_case_string(&self) -> Option<String> {
+        use std::fmt::Write;
+        use crate::traits::Message;
+
+        let mut s = String::new();
+
+        writeln!(s, "test CMSG_QUESTGIVER_ACCEPT_QUEST {{").unwrap();
+        // Members
+        writeln!(s, "    guid = {};", self.guid.guid()).unwrap();
+        writeln!(s, "    quest_id = {};", self.quest_id).unwrap();
+        writeln!(s, "    unk1 = {};", self.unk1).unwrap();
+
+        writeln!(s, "}} [").unwrap();
+
+        let [a, b] = 20_u16.to_be_bytes();
+        writeln!(s, "    {a:#04X}, {b:#04X}, /* size */").unwrap();
+        let [a, b, c, d] = 393_u32.to_le_bytes();
+        writeln!(s, "    {a:#04X}, {b:#04X}, {c:#04X}, {d:#04X}, /* opcode */").unwrap();
+        let mut bytes: Vec<u8> = Vec::new();
+        self.write_into_vec(&mut bytes).unwrap();
+        let mut bytes = bytes.into_iter();
+
+        crate::util::write_bytes(&mut s, &mut bytes, 8, "guid", "    ");
+        crate::util::write_bytes(&mut s, &mut bytes, 4, "quest_id", "    ");
+        crate::util::write_bytes(&mut s, &mut bytes, 4, "unk1", "    ");
+
+
+        writeln!(s, "] {{").unwrap();
+        writeln!(s, "    versions = \"{}\";", std::env::var("WOWM_TEST_CASE_WORLD_VERSION").unwrap_or("3.3.5".to_string())).unwrap();
+        writeln!(s, "}}\n").unwrap();
+
+        Some(s)
+    }
+
+    fn size_without_header(&self) -> u32 {
+        16
+    }
+
+    fn write_into_vec(&self, mut w: impl Write) -> Result<(), std::io::Error> {
+        // guid: Guid
+        w.write_all(&self.guid.guid().to_le_bytes())?;
+
+        // quest_id: u32
+        w.write_all(&self.quest_id.to_le_bytes())?;
+
+        // unk1: u32
+        w.write_all(&self.unk1.to_le_bytes())?;
+
+        Ok(())
+    }
+
+    fn read_body<S: crate::private::Sealed>(r: &mut &[u8], body_size: u32) -> Result<Self, crate::errors::ParseError> {
+        Self::read_inner(r, body_size).map_err(|a| crate::errors::ParseError::new(393, "CMSG_QUESTGIVER_ACCEPT_QUEST", body_size, a))
+    }
+
+}
+
+#[cfg(feature = "wrath")]
+impl crate::wrath::ClientMessage for CMSG_QUESTGIVER_ACCEPT_QUEST {}
+
