+++ wow_world_messages/src/world/wrath/mod.rs
@@ -697,7 +697,8 @@ pub use cmsg_query_inspect_achievements::*;
 pub(crate) mod cmsg_query_quests_completed;
 pub use cmsg_query_quests_completed::*;
 pub use crate::shared::cmsg_query_time_vanilla_tbc_wrath::*;
-pub use crate::shared::cmsg_questgiver_accept_quest_vanilla_tbc_wrath::*;
+pub(crate) mod cmsg_questgiver_accept_quest;
+pub use cmsg_questgiver_accept_quest::*;
 pub use crate::shared::cmsg_questgiver_cancel_vanilla_tbc_wrath::*;
 pub use crate::shared::cmsg_questgiver_choose_reward_vanilla_tbc_wrath::*;
 pub use crate::shared::cmsg_questgiver_complete_quest_vanilla_tbc_wrath::*;
