+++ wowm_language/src/docs/cmsg_questgiver_accept_quest.md
@@ -1,10 +1,10 @@
 # CMSG_QUESTGIVER_ACCEPT_QUEST
 
-## Client Version 1, Client Version 2, Client Version 3
+## Client Version 1, Client Version 2
 
 ### Wowm Representation
 
-Autogenerated from `wowm` file at [`wow_message_parser/wowm/world/quest/cmsg_questgiver_accept_quest.wowm:3`](https://github.com/gtker/wow_messages/tree/main/wow_message_parser/wowm/world/quest/cmsg_questgiver_accept_quest.wowm#L3).
+Autogenerated from `wowm` file at [`wow_message_parser/wowm/world/quest/cmsg_questgiver_accept_quest.wowm:1`](https://github.com/gtker/wow_messages/tree/main/wow_message_parser/wowm/world/quest/cmsg_questgiver_accept_quest.wowm#L1).
 ```rust,ignore
 cmsg CMSG_QUESTGIVER_ACCEPT_QUEST = 0x0189 {
     Guid guid;
@@ -28,4 +28,37 @@ CMSG have a header of 6 bytes.
 | ------ | ----------------- | ---- | ---- | ------- |
 | 0x06 | 8 / Little | [Guid](../types/packed-guid.md) | guid |  |
 | 0x0E | 4 / Little | u32 | quest_id |  |
+
+## Client Version 3.3.5
+
+The 3.3.5a client appends a trailing u32 that AzerothCore reads as `unk1`.
+
+### Wowm Representation
+
+Autogenerated from `wowm` file at [`wow_message_parser/wowm/world/quest/cmsg_questgiver_accept_quest.wowm:9`](https://github.com/gtker/wow_messages/tree/main/wow_message_parser/wowm/world/quest/cmsg_questgiver_accept_quest.wowm#L9).
+```rust,ignore
+cmsg CMSG_QUESTGIVER_ACCEPT_QUEST = 0x0189 {
+    Guid guid;
+    u32 quest_id;
+    u32 unk1;
+}
+```
+### Header
+
+CMSG have a header of 6 bytes.
+
+#### CMSG Header
+
+| Offset | Size / Endianness | Type   | Name   | Description |
+| ------ | ----------------- | ------ | ------ | ----------- |
+| 0x00   | 2 / Big           | uint16 | size   | Size of the rest of the message including the opcode field but not including the size field.|
+| 0x02   | 4 / Little        | uint32 | opcode | Opcode that determines which fields the message contains.|
+
+### Body
+
+| Offset | Size / Endianness | Type | Name | Comment |
+| ------ | ----------------- | ---- | ---- | ------- |
+| 0x06 | 8 / Little | [Guid](../types/packed-guid.md) | guid |  |
+| 0x0E | 4 / Little | u32 | quest_id |  |
+| 0x12 | 4 / Little | u32 | unk1 |  |
